+++ src/main/java/de/jeff_media/angelchest/hooks/IExecutableItemsHook.java
@@ -1,0 +1,10 @@
+package de.jeff_media.angelchest.hooks;
+
+import org.bukkit.inventory.ItemStack;
+
+public class IExecutableItemsHook {
+
+    public boolean isKeptOnDeath(ItemStack item) {
+        return false;
+    }
+}